backend/requirements.txt
flask
flask-cors
pymongo
firebase-admin
python-dotenv
reportlab
mistralai
